backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
pillow
python-dotenv
openai
